move DynamicType
