+++ src/private.rs
@@ -3,6 +3,7 @@ use crate::types::{
     author_replay::AuthorReplay,
     brush::{Brush, BrushGrid, BrushObject},
     color::Colour,
+    dynamic_type::DynamicType,
     exolvl::Exolvl,
     image::Image,
     layer::Layer,
@@ -63,7 +64,7 @@ impl_sealed!(
     super::Action,
     ActionType,
     super::NovaValue,
-    super::DynamicType,
+    DynamicType,
     super::FunctionCall,
     super::CallParameter,
     super::Variable,
